+++ PowerAdapterBlogs/settings/base.py
@@ -49,7 +49,7 @@ INSTALLED_APPS = [
     "jazzmin",
     "dal",
     "dal_select2",
-    "django.contrib.admin",
+    "PowerAdapterBlogs.admin_config.SuperuserAdminConfig",
     "django.contrib.auth",
     "django.contrib.contenttypes",
     "django.contrib.sessions",